remove polygon crypto service (2 separate fetchers extracted)
