+++ src/services/fetchers/common/PolygonIOStockPriceService.ts
@@ -2,13 +2,13 @@ import {inject, injectable} from 'inversify';
 import schedule, {Job} from 'node-schedule';
 import {Logger} from 'winston';
 
-import PolygonIOStockSnapshotFetcher from './fetchers/PolygonIOStockSnapshotFetcher.js';
-import PolygonIOSingleStockPriceFetcher from './fetchers/PolygonIOSingleStockPriceFetcher.js';
-import {SinglePriceResponse} from './fetchers/common/BasePolygonIOSingleFetcher.js';
-import {SnapshotResponse} from './fetchers/common/BasePolygonIOSnapshotFetcher.js';
-import PriceAggregator from './PriceAggregator.js';
-import Settings from '../types/Settings.js';
-import TimeService from './TimeService.js';
+import PolygonIOStockSnapshotFetcher from '../PolygonIOStockSnapshotFetcher.js';
+import PolygonIOSingleStockPriceFetcher from '../PolygonIOSingleStockPriceFetcher.js';
+import {SinglePriceResponse} from './BasePolygonIOSingleFetcher.js';
+import {SnapshotResponse} from './BasePolygonIOSnapshotFetcher.js';
+import PriceAggregator from '../../PriceAggregator.js';
+import Settings from '../../../types/Settings.js';
+import TimeService from '../../TimeService.js';
 
 @injectable()
 class PolygonIOStockPriceService {
