+++ src/services/fetchers/index.ts
@@ -1,4 +1,3 @@
-import PolygonIOCryptoPriceFetcher from './PolygonIOCryptoPriceFetcher.js';
 import PolygonIOStockPriceFetcher from './PolygonIOStockPriceFetcher.js';
 import PolygonIOCurrencySnapshotGramsFetcher from './PolygonIOCurrencySnapshotGramsFetcher.js';
 import YearnVaultTokenPriceFetcher from './YearnVaultTokenPriceFetcher.js';
@@ -21,7 +20,6 @@ export {
   PolygonIOCryptoSnapshotPriceFetcher,
   PolygonIOSingleCryptoPriceFetcher,
   PolygonIOStockPriceFetcher,
-  PolygonIOCryptoPriceFetcher,
   PolygonIOCurrencySnapshotGramsFetcher,
   YearnVaultTokenPriceFetcher,
   CoingeckoPriceFetcher,
