+++ src/types/fetchers.ts
@@ -6,7 +6,6 @@ import {EvmTWAPGasPriceInputParams} from 'src/services/fetchers/EvmTWAPGasPriceF
 import {GoldApiPriceInputParams} from 'src/services/fetchers/GoldApiPriceFetcher';
 import {MetalPriceApiInputParams} from 'src/services/fetchers/MetalPriceApiFetcher';
 import {MetalsDevApiPriceInputParams} from 'src/services/fetchers/MetalsDevApiFetcher';
-import {PolygonIOCryptoPriceInputParams} from 'src/services/fetchers/PolygonIOCryptoPriceFetcher';
 import {PolygonIOCurrencySnapshotGramsInputParams} from 'src/services/fetchers/PolygonIOCurrencySnapshotGramsFetcher';
 import {PolygonIOPriceInputParams} from 'src/services/fetchers/PolygonIOStockPriceFetcher';
 import {PolygonIOCryptoSnapshotInputParams} from '../services/fetchers/PolygonIOCryptoSnapshotPriceFetcher.js';
@@ -37,7 +36,6 @@ export type FeedFetcherInputParams =
   | GoldApiPriceInputParams
   | PolygonIOCryptoSnapshotInputParams[]
   | PolygonIOSingleCryptoPriceInputParams[]
-  | PolygonIOCryptoPriceInputParams
   | PolygonIOCurrencySnapshotGramsInputParams
   | PolygonIOPriceInputParams
   | EvmTWAPGasPriceInputParams
